examples/requirements.txt
aiohttp-tal
aiohttp_babel
babel-lingua-chameleon